+++ lib/dart_husky.dart
@@ -3,3 +3,5 @@ library;
 
 export 'src/config/config_model.dart';
 export 'src/hooks/commit_msg_validator.dart';
+export 'src/utils/melos_utils.dart';
+export 'src/utils/process_utils.dart';